+++ R/as_period.R
@@ -17,8 +17,6 @@
 #'
 #' This function respects [dplyr::group_by()] groups.
 #'
-#' Currently periods finer than second data are not supported.
-#'
 #' The `side` argument is useful when you want to return data at, say, the
 #' end of a quarter, or the end of a month.
 #'
@@ -77,6 +75,9 @@
 #'
 #' # Using start_date ----------------------------------------------------------
 #'
+#'
+#' #### One method using start_date
+#'
 #' # FB stock prices
 #' data(FB)
 #' FB <- as_tbl_time(FB, date)
@@ -89,6 +90,22 @@
 #' # Specifying the `start_date = "2013-01-01"` might be preferable.
 #' # Groups become (2013-01-01, 2013-01-02), (2013-01-03, 2013-01-04) and so on.
 #' as_period(FB, "2 day", start_date = "2013-01-01")
+#'
+#' #### Equivalent method using an index vector
+#'
+#' # FB stock prices
+#' data(FB)
+#' FB <- as_tbl_time(FB, date)
+#'
+#' custom_period <- create_series(
+#'   time_formula = dplyr::first(FB$date) - 1 ~ dplyr::last(FB$date),
+#'   period       = "2 day",
+#'   class        = "Date",
+#'   as_vector    = TRUE)
+#'
+#' FB %>%
+#'   as_tbl_time(date) %>%
+#'   as_period(period = custom_period)
 #'
 #' # Manually calculating returns at different periods -------------------------
 #'