+++ R/compat-tidyr.R
@@ -30,7 +30,9 @@ nest.tbl_time <- function(data, ..., .key = "data") {
 
 unnest.tbl_time <- function(data, ..., .drop = NA, .id = NULL, .sep = NULL) {
   # This is called after nesting but excluding the index in the nest
-  reconstruct(NextMethod(), data)
+  #reconstruct(NextMethod(), data)
+  copy_.data <- new_tbl_time(data, get_index_quo(data), get_index_time_zone(data))
+  reconstruct(NextMethod(), copy_.data)
 }
 
 unnest.tbl_df <- function(data, ..., .drop = NA, .id = NULL, .sep = NULL) {
@@ -74,36 +76,30 @@ unnest.tbl_df <- function(data, ..., .drop = NA, .id = NULL, .sep = NULL) {
 
 
 # ------------------------------------------------------------------------------
-# Unsure if gather / spread are needed, but these would do the trick
+# gather() and spread() seem to be needed as well
 
-# #' @export
-# #' @importFrom tidyr gather
-# #'
-# gather.tbl_time <- function(data, key = "key", value = "value", ..., na.rm = FALSE,
-#                             convert = FALSE, factor_key = FALSE)  {
-#
-#   key   <- rlang::enquo(key)
-#   value <- rlang::enquo(value)
-#   quos  <- rlang::quos(...)
-#
-#   gathered_data <- gather(as_tibble(data), key = !! key, value = !! value, !!! quos,
-#                           na.rm = na.rm, convert = convert, factor_key = factor_key)
-#
-#   reconstruct(gathered_data, data)
-# }
+gather.tbl_time <- function(data, key = "key", value = "value", ..., na.rm = FALSE,
+                            convert = FALSE, factor_key = FALSE)  {
 
-# #' @export
-# #' @importFrom tidyr spread
-# #'
-# spread.tbl_time <- function(data, key, value, fill = NA, convert = FALSE, drop = TRUE,
-#                             sep = NULL)  {
-#
-#   key   <- rlang::enquo(key)
-#   value <- rlang::enquo(value)
-#
-#   spread_data <- spread(as_tibble(data), key = !! key, value = !! value,
-#                         fill = fill, convert = convert, drop = drop,
-#                         sep = sep)
-#
-#   reconstruct(spread_data, data)
-# }
+  key   <- rlang::enquo(key)
+  value <- rlang::enquo(value)
+  quos  <- rlang::quos(...)
+
+  gathered_data <- tidyr::gather(as_tibble(data), key = !! key, value = !! value, !!! quos,
+                          na.rm = na.rm, convert = convert, factor_key = factor_key)
+
+  reconstruct(gathered_data, data)
+}
+
+spread.tbl_time <- function(data, key, value, fill = NA, convert = FALSE, drop = TRUE,
+                            sep = NULL)  {
+
+  key   <- rlang::enquo(key)
+  value <- rlang::enquo(value)
+
+  spread_data <- tidyr::spread(as_tibble(data), key = !! key, value = !! value,
+                        fill = fill, convert = convert, drop = drop,
+                        sep = sep)
+
+  reconstruct(spread_data, data)
+}